refactor: move Combo struct to another file
